Move TransferLearnConfig to types.ts; export it
